+++ loom/__init__.py
@@ -1,2 +1,2 @@
 """loom/__init__.py"""
-__version__ = "0.6.1"
+__version__ = "0.7.0"